Move test data to test folder
