+++ product_code/authentication_service/api/src/test/scala/de/upb/cs/uc4/authentication/DefaultTestAuthenticationUsers.scala
@@ -1,4 +1,4 @@
-package de.upb.cs.uc4.authentication.test
+package de.upb.cs.uc4.authentication
 
 import de.upb.cs.uc4.authentication.model.{ AuthenticationRole, AuthenticationUser }
